fixed air temp target slider behavior
