+++ client/src/components/StageTabs/AirTemperatureSelectorFunctional.js
@@ -1,0 +1,87 @@
+/*
+ * Copyright (c) John Rodley 2022.
+ * SPDX-FileCopyrightText:  John Rodley 2022.
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import React, { useState } from 'react';
+
+import { Grid,Image, Box, RangeInput } from 'grommet';
+import thermometer from '../../images/thermometer-icon.png'
+import './stagesTab.css';
+import '../../Palette.css';
+
+// copyright and license inspection - no issues 4/13/22
+
+function RenderAirTemperatureSelector(props) {
+    function setValue(value) {
+        let x = props.automation_setting;
+        x.target_temperature=value;
+        props.setAutomationSettingFromChild(x)
+    }
+    const changeTarget = event => setValue(event.target.value);
+    const units = props.display_settings.temperature_units;
+    const min = props.automation_setting.temperature_min;
+    const max = props.automation_setting.temperature_max;
+
+    const target_temperature = props.automation_setting.target_temperature
+
+    return (
+        <Grid
+            round={'xxsmall'}
+            fill
+            areas={[
+                { name: 'label', start: [0, 0], end: [3, 0] },
+                { name: 'icon', start: [0, 1], end: [0, 1] },
+                { name: 'min', start: [1, 1], end: [1, 1] },
+                { name: 'slider', start: [2, 1], end: [2, 1] },
+                { name: 'max', start: [3, 1], end: [3, 1] },
+                { name: 'value', start: [0, 2], end: [3, 2] },
+            ]}
+            columns={['16px', '30px', '400px', '30px']}
+            rows={['28px','60px', '28px']}
+            gap="xxsmall"
+        >
+            <Box gridArea="label" color={'yellow'}  align="center">
+                {props.label}
+            </Box>
+
+            <Box gridArea="icon"  justify={"center"}>
+                <Image src={thermometer} />
+            </Box>
+            <Box gridArea="min" justify={"center"}>{min}{units}</Box>
+            <Box gridArea="slider"  justify={"center"} align-content={'center'}>
+                <RangeInput
+                    direction="horizontal"
+                    min={min}
+                    max={max}
+                    step={1}
+                    value={target_temperature}
+                    onChange={changeTarget}
+                />
+            </Box>
+            <Box gridArea="max" justify={"center"}>{max}{units}</Box>
+            <Box gridArea="value" justify={"center"} align="center">
+                {target_temperature}{units}
+            </Box>
+        </Grid>
+    );
+}
+export default RenderAirTemperatureSelector;
